+++ src/main/java/gtclassic/common/GTIcons.java
@@ -85,6 +85,7 @@ public class GTIcons {
 		setTexture(GTBlocks.tileRedstoneTransmitter, 0, 64, 2, 24, 2, 2, 0, 64, 27, 27, 2, 2);
 		setTexture(GTBlocks.tileRedstoneReceiver, 24, 27);
 		setTexture(GTBlocks.tileBlockExtender, 96, 97, 55, 98, 98, 98);
+		setTexture(GTBlocks.tileRockBreaker, 53, 53, 53, 61, 53, 53);
 	}
 
 	/**